+++ resources/views/components/message.blade.php
@@ -4,7 +4,7 @@
     'class' => '',
 ])
 
-<div class="toast text-white bg-{{ $type }} m-3 border-0 {{ $class ?? '' }}" role="alert" aria-live="assertive" aria-atomic="true"
+<div class="toast text-white bg-{{ $type == 'error' ? 'danger' : $type }} m-3 border-0 {{ $class ?? '' }}" role="alert" aria-live="assertive" aria-atomic="true"
     data-bs-autohide="{{ $type == 'success' ? 'true' : 'false' }}"
     data-bs-delay="8000"
     style="