+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.11.0] - 2025-04-16
+
+_Stable release based on [0.11.0-rc.1]._
+
 ## [0.11.0-rc.1] - 2025-04-16
 
 ### Changed
@@ -316,6 +320,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[0.11.0]: https://https://github.com/internetguru/laravel-common/compare/v0.10.2...v0.11.0
 [0.11.0-rc.1]: https://github.com/internetguru/laravel-common/releases/tag/v0.10.2
 [0.10.2]: https://https://github.com/internetguru/laravel-common/compare/v0.10.1...v0.10.2
 [0.10.1]: https://https://github.com/internetguru/laravel-common/compare/v0.10.0...v0.10.1